+++ CHANGELOG.md
@@ -13,9 +13,10 @@
 * replace `ParameterOrRef` with `ReferenceOr<Parameter>`
 * rename `Spec` to `OpenAPI` to match openapiv3
 * replace `Option<Vec<T>>` with `Vec<T>` and skip if `is_empty` like openapiv3
+* add dependency on `indexmap` like openapiv3
+* use `is_empty` on `IndexMap` instead of wrapping in `Option<T>` like openapiv3
 
 * expose security definition as an enum type
-* expose schema types which define `parameters` with new `ParameterOrRef` type
 * Adds License object
 * Adds Contact object
 * Derives Default for all structs